LJS-119-dag storybook upgraded
