--- packages/dag/.storybook/.babelrc
@@ -1,3 +1,0 @@
-{
-  "extends": "../.babelrc"
-}
