+++ packages/dag/.storybook/preview.js
@@ -1,0 +1,4 @@
+
+export const parameters = {
+  actions: { argTypesRegex: "^on[A-Z].*" },
+}
