--- packages/dag/.storybook/webpack.config.js
@@ -1,15 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  module: {
-    rules: [
-      {
-        test: /\.jsx?$/,
-        include: path.resolve('./stories'),
-        loader: require.resolve('@storybook/addon-storysource/loader'),
-        options: { prettierConfig: { parser: 'babel' } },
-        enforce: 'pre'
-      }
-    ]
-  }
-};
