+++ packages/dag/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@latticejs/dag",
-  "version": "1.0.1-beta.2",
+  "version": "1.0.1-beta.4",
   "license": "MIT",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
@@ -10,17 +10,18 @@
   ],
   "scripts": {
     "build:clean": "rimraf dist/**",
-    "build": "npm run build:clean && lattice-scripts build ./src/index.js",
-    "build:dev": "npm run build:clean && lattice-scripts build -e development ./src/index.js",
+    "build": "npm run build:clean && lattice-scripts build ./src/components/index.js",
+    "build:dev": "npm run build:clean && lattice-scripts build -e development ./src/components/index.js",
     "release": "npm run build",
     "prepublish": "npm run release",
-    "storybook": "start-storybook -p 9010 -c .storybook",
+    "storybook": "start-storybook -p 6006",
     "test": "jest",
-    "size": "bundlesize"
+    "size": "bundlesize",
+    "build-storybook": "build-storybook"
   },
   "peerDependencies": {
-    "@material-ui/core": "^4.8.2",
-    "@material-ui/icons": "^4.5.1",
+    "@material-ui/core": "^5.8.2",
+    "@material-ui/icons": "^5.5.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   },
@@ -35,11 +36,13 @@
     "@latticejs/storybook-readme": "^1.0.1-beta.1",
     "@material-ui/core": "^4.8.2",
     "@material-ui/icons": "^4.5.1",
-    "@sambego/storybook-state": "^1.3.0",
-    "@storybook/addon-actions": "^5.3.19",
-    "@storybook/addon-storyshots": "^5.3.19",
-    "@storybook/addon-storysource": "^5.3.19",
-    "@storybook/react": "^5.3.19",
+    "@sambego/storybook-state": "^2.0.1",
+    "@storybook/addon-actions": "^6.0.26",
+    "@storybook/addon-essentials": "^6.0.26",
+    "@storybook/addon-links": "^6.0.26",
+    "@storybook/addon-storyshots": "^6.0.26",
+    "@storybook/addon-storysource": "^6.0.26",
+    "@storybook/react": "^6.0.26",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^24.9.0",
@@ -52,10 +55,11 @@
     "npm-run-all": "^4.1.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-is": "^16.13.1",
     "react-jss": "^10.0.0",
     "react-test-renderer": "^16.13.1",
     "rimraf": "^3.0.0",
-    "storybook-readme": "^4.0.5"
+    "storybook-readme": "^5.0.8"
   },
   "dependencies": {
     "classnames": "^2.2.0",
